chore:#04 app.jsno のコード記述順番を整理
