+++ src/post-processing/app.js
@@ -2,89 +2,50 @@
  * Three.js
  * https://threejs.org/
  */
-import * as THREE from "three";
-import vertexShader from "./vertex.glsl";
-import fragmentShader from "./fragment.glsl";
-import GUI from "lil-gui";
-import { gsap } from "gsap";
-import { OrbitControls } from "three/examples/jsm/controls/OrbitControls";
+import * as THREE from 'three';
+import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls';
 
-init();
-async function init() {
-  const scene = new THREE.Scene();
-  const camera = new THREE.PerspectiveCamera(
-    75,
-    window.innerWidth / window.innerHeight,
-    0.1,
-    1000
-  );
+// メインのレンダラーの設定
+const renderer = new THREE.WebGLRenderer({ antialias: true });
+renderer.setSize(window.innerWidth, window.innerHeight);
+renderer.setClearColor(0xeeeeee);
+document.body.appendChild(renderer.domElement);
 
-  const renderer = new THREE.WebGLRenderer({ antialias: true });
-  renderer.setSize(window.innerWidth, window.innerHeight);
-  renderer.setClearColor(0xffffff);
-  document.body.appendChild(renderer.domElement);
+const camera = new THREE.PerspectiveCamera(75, window.innerWidth / window.innerHeight, 0.1, 1000);
+camera.position.set(5,5,5);
 
-  async function loadTex(url) {
-    const texLoader = new THREE.TextureLoader();
-    const texture = await texLoader.loadAsync(url);
-    texture.wrapS = THREE.ClampToEdgeWrapping;
-    texture.wrapT = THREE.MirroredRepeatWrapping;
-    return texture;
-  }
+const controls = new OrbitControls(camera, renderer.domElement);
 
-  const geometry = new THREE.PlaneGeometry(25, 12.5);
-  const material = new THREE.ShaderMaterial({
-    uniforms: {
-      uTex1: { value: await loadTex("/img/output3.jpg") },
-      uTex2: { value: await loadTex("/img/output2.jpg") },
-      uTick: { value: 0 },
-      uProgress: { value: 0 },
-    },
-    vertexShader,
-    fragmentShader,
-  });
-  const plane = new THREE.Mesh(geometry, material);
-  scene.add(plane);
+const rtGeo = new THREE.BoxGeometry(4, 4, 4);
+const rtMate = new THREE.MeshLambertMaterial({ color: 0x009dff, side: THREE.DoubleSide });
+const rtMesh = new THREE.Mesh(rtGeo, rtMate);
+const scene = new THREE.Scene();
+scene.background = new THREE.Color( 0xeeeeee );
+const light1 = new THREE.PointLight( 0xffffff, 1, 0 );
+light1.position.set( 0, 20, 0 );
+scene.add( light1 );
 
-  camera.position.z = 30;
+const light2 = new THREE.PointLight( 0xffffff, 1, 0 );
+light2.position.set( 10, 20, 10 );
+scene.add( light2 );
 
-  const axis = new THREE.AxesHelper(100);
-  scene.add(axis);
+const light3 = new THREE.PointLight( 0xffffff, 1, 0 );
+light3.position.set( - 10, - 20, - 10 );
+scene.add( light3 );
 
-  const controls = new OrbitControls(camera, renderer.domElement);
-  controls.enableDamping = true;
+scene.add(rtMesh);
 
-  const gui = new GUI();
-  const folder1 = gui.addFolder("z-distance");
-  folder1.open();
 
-  folder1
-    .add(material.uniforms.uProgress, "value", 0, 1, 0.1)
-    .name("zaxis")
-    .listen();
+function animate() {
+  requestAnimationFrame(animate);
 
-  const datData = { next: !!material.uniforms.uProgress.value };
-  folder1
-    .add(datData, "next")
-    .name("moving axis")
-    .onChange(() => {
-      gsap.to(material.uniforms.uProgress, {
-        value: datData.next ? 1 : 0,
-        duration: 3,
-        ease: "ease",
-      });
-    });
+  // メインのレンダラーの描写
+  renderer.render(scene, camera);
 
-  let i = 0;
-  function animate() {
-    requestAnimationFrame(animate);
-    controls.update();
+  rtMesh.rotation.x += 0.01;
+  rtMesh.rotation.y += 0.01;
 
-    // cube.rotation.x = cube.rotation.x + 0.01;
-    // cube.rotation.y += 0.01;
-
-    renderer.render(scene, camera);
-  }
-
-  animate();
+  controls.update();
 }
+
+animate();
