+++ src/post-processing/bk_app.js
@@ -1,0 +1,95 @@
+/**
+ * Three.js
+ * https://threejs.org/
+ */
+import * as THREE from "three";
+import vertexShader from "./vertex.glsl";
+import fragmentShader from "./fragment.glsl";
+import GUI from "lil-gui";
+import { gsap } from "gsap";
+import { OrbitControls } from "three/examples/jsm/controls/OrbitControls";
+
+init();
+async function init() {
+  //メインのレンダラー
+  const renderer = new THREE.WebGLRenderer({ antialias: true });
+  renderer.setSize(window.innerWidth, window.innerHeight);
+  renderer.setClearColor(0xffffff);
+  document.body.appendChild(renderer.domElement);
+  const scene = new THREE.Scene();
+
+
+  const camera = new THREE.PerspectiveCamera(
+    75,
+    window.innerWidth / window.innerHeight,
+    0.1,
+    1000
+  );
+  camera.position.z = 30;//set()methodにあとで変更する
+
+  const controls = new OrbitControls(camera, renderer.domElement);
+  controls.enableDamping = true;
+
+  async function loadTex(url) {
+    const texLoader = new THREE.TextureLoader();
+    const texture = await texLoader.loadAsync(url);
+    texture.wrapS = THREE.ClampToEdgeWrapping;
+    texture.wrapT = THREE.MirroredRepeatWrapping;
+    return texture;
+  }
+
+  const geometry = new THREE.PlaneGeometry(25, 12.5);
+  const material = new THREE.ShaderMaterial({
+    uniforms: {
+      uTex1: { value: await loadTex("/img/output3.jpg") },
+      uTex2: { value: await loadTex("/img/output2.jpg") },
+      uTick: { value: 0 },
+      uProgress: { value: 0 },
+    },
+    vertexShader,
+    fragmentShader,
+  });
+  const mesh = new THREE.Mesh(geometry, material);
+  scene.add(mesh);
+
+
+
+  const axis = new THREE.AxesHelper(100);
+  scene.add(axis);
+
+  
+
+  const gui = new GUI();
+  const folder1 = gui.addFolder("z-distance");
+  folder1.open();
+
+  folder1
+    .add(material.uniforms.uProgress, "value", 0, 1, 0.1)
+    .name("zaxis")
+    .listen();
+
+  const datData = { next: !!material.uniforms.uProgress.value };
+  folder1
+    .add(datData, "next")
+    .name("moving axis")
+    .onChange(() => {
+      gsap.to(material.uniforms.uProgress, {
+        value: datData.next ? 1 : 0,
+        duration: 3,
+        ease: "ease",
+      });
+    });
+
+  let i = 0;
+  function animate() {
+    requestAnimationFrame(animate);
+    controls.update();
+
+    // cube.rotation.x = cube.rotation.x + 0.01;
+    // cube.rotation.y += 0.01;
+
+    renderer.render(scene, camera);
+  }
+
+  animate();
+}
